requirements.txt
Pillow >= 10.0, < 11.0
numpy >= 1.0, < 2.0
discord >= 2.0, < 3.0
attrs >= 23.0, < 24.0
